refactor(frontend): split visible cell runtime
